agents/ecommerce-domain-agent/service/src/ecommerce_agent/clients/test_case_generator.py
"""Client for Test Case Generator agent."""

import grpc
from typing import Any, Dict
from ecommerce_agent.config import get_settings
from ecommerce_agent.utils.logging import get_logger

logger = get_logger(__name__)

class TestCaseGeneratorClient:
    """Client for communicating with Test Case Generator agent."""

    def __init__(self):
        config = get_settings()
        # Test Case Generator would run on a different port
        self.host = config.test_case_generator_host if hasattr(config, 'test_case_generator_host') else "localhost"
        self.port = config.test_case_generator_port if hasattr(config, 'test_case_generator_port') else 9093

    async def generate_test_cases(self, request: Dict[str, Any]) -> Dict[str, Any]:
        """
        Generate test cases via Test Case Generator agent.

        Args:
            request: Test case generation request

        Returns:
            Generated test cases
        """
        # For now, return simulated response
        # In production, this would make actual gRPC call
        logger.info(
            "Generating test cases (simulated)",
            entity=request.get("entity"),
            workflow=request.get("workflow"),
            method=request.get("generation_method")
        )

        # Simulate test case generation
        test_cases = []
        count = request.get("count", 10)
        entity = request.get("entity")
        test_types = request.get("test_types", ["functional"])

        for i in range(count):
            test_case = {
                "id": f"TC-{entity.upper()}-{i+1:04d}",
                "name": f"Test {entity} {test_types[0]} scenario {i+1}",
                "description": f"Validate {entity} {test_types[0]} behavior in scenario {i+1}",
                "type": test_types[0],
                "priority": "high" if i < 3 else "medium",
                "preconditions": [
                    f"System is in valid state",
                    f"{entity} entity exists"
                ],
                "steps": [
                    {"action": f"Initialize {entity}", "expected_result": f"{entity} initialized successfully"},
                    {"action": f"Perform {test_types[0]} operation", "expected_result": "Operation succeeds"},
                    {"action": "Validate result", "expected_result": "Result matches expectations"}
                ],
                "postconditions": ["System state is consistent"],
                "test_data": {
                    "entity": entity,
                    "sample_data": f"test_data_{i+1}"
                },
                "business_rules": request.get("domain_context", {}).get("business_rules", [])[:2],
                "expected_behavior": f"{entity} behaves according to business rules",
                "edge_cases": request.get("domain_context", {}).get("edge_cases", [])[:1] if request.get("include_edge_cases") else [],
                "automation_feasibility": True,
                "estimated_duration": 5 + (i % 3) * 2
            }
            test_cases.append(test_case)

        return {
            "suite_id": f"TS-{entity.upper()}-001",
            "suite_name": f"{entity} Test Suite",
            "test_cases": test_cases,
            "coverage_percentage": 85.0 + (count * 0.5),
            "business_context": {
                "entity": entity,
                "domain": "ecommerce",
                "test_types": test_types
            }
        }

    async def health_check(self) -> bool:
        """Check if Test Case Generator agent is healthy."""
        try:
            # In production, make actual health check call
            return True
        except Exception as e:
            logger.error(
                "Test Case Generator health check failed",
                error=str(e)
            )
            return False